naruby/bench+docs: 値型を int64_t に統一し仕様/実装を分離
- bench/*.c: int → int64_t に統一 (naruby は VALUE=int64_t、C 側は
  混在で一部 int32 overflow していた)。perf.md フルマトリクスの
  gcc 列を再計測。branch_dom の int32 overflow 注記を撤去。
- spec.md: 言語仕様のみに絞る (parser/callcache/RESULT 等の実装は
  runtime.md に集約)。264 → 約 190 行。
- perf.md: 「naruby vs C 仕様比較」セクションを追加。late binding
  (callcache 検査) を主因として明示。yjit 比較テーブルから yjit
  列を除去 (フルマトリクスでは参考値として残す)。

+++ sample/naruby/bench/branch_dom.c
@@ -1,8 +1,9 @@
 #include <stdio.h>
+#include <stdint.h>
 
 #define A __attribute__((noinline,noipa))
 
-static A int f(int n) {
+static A int64_t f(int64_t n) {
     if (n < 1000000000) {
         return n + 1;
     } else {
@@ -12,11 +13,11 @@ static A int f(int n) {
 
 int main(void)
 {
-    int sz = 50 * 1000 * 1000;
-    int acc = 0;
-    for (int i = 0; i < sz; i++) {
+    int64_t sz = 50 * 1000 * 1000;
+    int64_t acc = 0;
+    for (int64_t i = 0; i < sz; i++) {
         acc += f(42);
     }
-    printf("%d\n", acc);
+    printf("%ld\n", (long)acc);
     return 0;
 }
